requirements.txt
numpy>=1.21.0
torch>=1.9.0
rasterio>=1.2.10
sentinelsat>=1.1.1
matplotlib>=3.4.3
scikit-learn>=0.24.2
tqdm>=4.62.3
pandas>=1.3.3
geopandas>=0.9.0
shapely>=1.7.1
shap>=0.40.0
pillow>=8.3.2
requests>=2.26.0 